pyproject.toml
[tool.poetry]
name = "wolf-catcher"
version = "0.1.0"
description = ""
authors = ["gyuha <nicegyuha@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
arrow = "^1.2.2"
bs4 = "^0.0.1"
PySide6 = "^6.3.1"
selenium = "^4.3.0"
Pillow = "^9.2.0"
requests = "^2.28.1"
PyYAML = "^6.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pyinstaller = "^5.3"
autopep8 = "^1.7.0"
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
